requirements.txt
django-mptt
djangorestframework
pillow
django-taggit
